Crate optee_utee_macros

Attribute Macros

ta_close_session

Attribute to declare the entry point of closing a session. Session context raw pointer (*mut T) can be defined as an optional parameter.

ta_create

Attribute to declare the entry point of creating TA.

ta_destroy

Attribute to declare the entry point of destroying TA.

ta_invoke_command

Attribute to declare the entry point of invoking commands. Session context reference (&mut T) can be defined as an optional parameter.

ta_open_session

Attribute to declare the entry point of opening a session. Pointer to session context pointer (*mut *mut T) can be defined as an optional parameter.